Added changesize, and pointerselection
Attach changesize to each new object, it will be activated when the object is selected, and disabled when the object is not selected. Pointerselection is a script that is attached to a pointer attached to the left hand, something I haven't done yet was to attach the pointer script directly to the oculus controller, but for now it is on a rectangle that is visible on the left hand for testing. 

BUTTONS: LEFT TRIGGER, RIGHT TRIGGER, SQUEEZE LEFT, SQUEEZE RIGHT.